修改内容： 1. [项目] 新增ClientTest项目，用于测试服务接口 2. [Database] 新增表TNickname 3. [Database] 修改TAccount表的GetKey函数 4. [Database] 修复保存数据时，错误码返回的错误 5. [Frame] 修复状态机 状态不迁移的问题 6. [Frame] 增加新连接加入和连接断开的回调 7. [GameServer] 新增创角接口 8. [DirServer] 目录服务器新增注册接口

+++ DataBase/Tables/TAccount.cs
@@ -12,12 +12,12 @@ namespace DataBase.Tables
 
         protected override string GetKey()
         {
-            return string.Format("{0}|{1}|{2}", Value.Type, Value.Account,Value.Zone);
+            return string.Format("{0}|{1}|{2}", ((int)Value.Type), Value.Account,Value.Zone);
         }
 
         protected override string GetKey((AuthType, string, int) key)
         {
-            return string.Format("{0}|{1}|{2}", key.Item1, key.Item2, key.Item3);
+            return string.Format("{0}|{1}|{2}", ((int)key.Item1), key.Item2, key.Item3);
         }
     }
 }
